requirements.txt
whois
pillow
cryptography